Extract bookstore actions into a separate file
